+++ pkg/microservice/aslan/core/environment/service/product.go
@@ -34,6 +34,7 @@ import (
 	"github.com/koderover/zadig/pkg/setting"
 	e "github.com/koderover/zadig/pkg/tool/errors"
 	"github.com/koderover/zadig/pkg/tool/log"
+	"github.com/koderover/zadig/pkg/util"
 )
 
 var DefaultCleanWhiteList = []string{"spockadmin"}
@@ -144,15 +145,11 @@ func GetInitProduct(productTmplName string, log *zap.SugaredLogger) (*commonmode
 			if serviceTmpl.Type == setting.K8SDeployType || serviceTmpl.Type == setting.HelmDeployType {
 				serviceResp.Containers = make([]*commonmodels.Container, 0)
 				for _, c := range serviceTmpl.Containers {
-					imageName := c.ImageName
-					if imageName == "" {
-						imageName = c.Name
-					}
 					container := &commonmodels.Container{
 						Name:      c.Name,
 						Image:     c.Image,
 						ImagePath: c.ImagePath,
-						ImageName: imageName,
+						ImageName: util.GetImageNameFromContainerInfo(c.ImageName, c.Name),
 					}
 					serviceResp.Containers = append(serviceResp.Containers, container)
 				}
