+++ pkg/util/yaml.go
@@ -42,15 +42,3 @@ func SplitManifests(content string) []string {
 
 	return res
 }
-
-// image ccr.ccs.tencentyun.com/koderover/nginx:stable
-// return nginx
-func GetImageName(image string) string {
-	imageNameStr := ""
-	imageArr := strings.Split(image, ":")
-	if len(imageArr) > 0 {
-		imageNameArr := strings.Split(imageArr[0], "/")
-		imageNameStr = imageNameArr[len(imageNameArr)-1]
-	}
-	return imageNameStr
-}
